refactor: Update TraineeList.jsx
